[fix]: deleteOldBestsellers 조건 수정 (#42)

+++ src/main/java/com/example/bookjourneybackend/domain/book/service/BestSellerService.java
@@ -113,8 +113,8 @@ public class BestSellerService {
 
         //기존 베스트 셀러 여부 바꾸기
         oldBestSeller.setBestSeller(false);
-        //이 책에해당하는 방이 없으면 db에서 삭제
-        if(oldBestSeller.getRooms().isEmpty()) {
+        //이 책에해당하는 방이 없거나, 이책을 즐겨찾기 한사람이 없으면 db에서 삭제
+        if(oldBestSeller.getRooms().isEmpty() && oldBestSeller.getFavorites().isEmpty()) {
             bookRepository.delete(oldBestSeller);
         }
     }
